Support for pre-packaged charts in frigate
If a user packs the chart properly (including its dependent charts)
so he would only like frigate to just go through them and just dump
the values from them that way it already does.
The only thing would be to NOT run the repo update and
the dep update commands.
One thing to note here is that the above commands only slow down
the whole recursion drastically (I had to wait for 20 mins).

With this improvement a flag called --no-update has been introduced
which would not invoke the the aforementioned commands
during the traversal.

Fixes #26